[getStats] Add outbound-rtp.simulcastIndex metric.
Helps RID-agnostic use cases to know which encoding have which SSRC,
including apps not using RIDs yet or stats processing that does not know
the app-specific mapping between RID and index such as
chrome://webrtc-internals/ having zero knowledge about app choices.

For now this is a C++-only metric, but we may try to standardize it.

Tested: Both unit test and manually in chrome using webrtc-internals
that for simulcast demos we get simulcastIndex and for singlecast demos
it is left undefined.

Bug: chromium:404853839
Change-Id: Idbb252526a1fca12c1bc837ad987c77fc17cb0fb
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/382200
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44196}