dcsctp: Add pull mode to consume received messages
The current API delivers received message by a callback API. The caller
is expected to process the received message directly, and the socket
will immediately restore the advertised receiver window as there is
nothing buffered in the socket any longer. But it may be buffered in the
application, if it can't process messages in the same rate as they are
received.

How can you as a receiver tell the sender to stop sending for a while,
because it can't keep up with the pace? It's not possible - the callback
is triggered unconditionally.

In other words: The advertised receiver window is of little use for
implementing back-pressure, which it's intending to do.

This CL adds a new option, that might become the default, that replaces
the delivery callback with a new "signal" callback, just indicating to
the caller that there are received messages that can be consumed by
calling a new API on the socket. If not, the message statys buffered and
it's reflected in the advertised receiver window.

Note that as a back-pressure algorithm, it's quite simple. The sender
must make sure that it doesn't send more than the receiver can fit
within its receiver window, but it's also allowed to always have a
single MTU inflight. So the sender will never stop sending completely,
but will definitely go down in send rate.

Depending on application, it might be a good idea to ALSO implement a
more sophisticated back-pressure algorithm by e.g. sending backpressure
messages on a separate stream that has a higher priority. But this
CL provides the fallback basics that the SCTP protocol provides.

No-Iwyu: Base CL not clean. Needs separate cleaning.
Co-authored-by: Nathan Lewis <npl@google.com>
Bug: webrtc:407622662
Change-Id: I9a63ee117a0fcf0ccb8409287b6a0548f9bcb7b1
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/383780
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Victor Boivie <boivie@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44486}