dtls-in-stun: Prepare for multi packet handshakes
1) Change format of ACK attribute to be hash of packet (crc32).
this is good since we anyway can't retransmit parts of DTLS packet
and it also works for the encrypted handshake packages.

This change is "somewhat upgrade" compatible,
version prior to this does not actually check/use the content of
the ack attribute, only the presence.

This version will automatically clear pending packets
when a new flight begins => It works, but retransmit of 1
packet in a multi placket flight does not unless both
ens has this version....But multi packet handshake flight
are not yet "in play"

2) Keep track of individually acked packets.
Clear array of pending when getting an ACK. This is "already today"
and improvment of existing code.

3) Limit size of ACK to 4 packets,
this limit *should* never be reached...but there is a test for it :)

4) Readd restrictions removed in https://webrtc-review.googlesource.com/c/src/+/381102. It only worked "by accident",
and the use case is of zero value. Even hard to explain!
Maybe I'll take a stab after PQC is done.

---

With this, PQC handshake _almost_ work with dtls-in-stun.

BUG=webrtc:367395350

Change-Id: I582397babb099896ba56ef26e174f47f7e98d1d9
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/385401
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44340}