Make `rtc_event_begin_end` publically visible.
This allows downstream users to use the
`ParsedRtcEventLog::{start,stop}_log_events()` accessors. This can, for example, be useful in unit tests where a final stop event is emitted to indicate the end of a log.

All other event targets in this build file are publicly visible, so this
is probably an oversight.

Change-Id: Iba17de1f5b91a297db939459cd78e5ad070d2d48
Bug: b/394338823
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/388060
Reviewed-by: Björn Terelius <terelius@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44455}