[iOS] Update RTCCameraPreviewView to conform to iOS 17
In iOS 17, the video orientation API is deprecated. This CL updates the
RTCCameraPreviewView file to use the video rotation angle API instead.

This file is being updated now to support Chrome for iOS effort to raise
its minimum deployment target to iOS 17.

In addition, this CL is a follow-up to the linked reverted CL that
fixes the incorrect build guard that gated the deprecated code.

Change-Id: I82feadbdca227259856948f0b683478ec52535cb
Bug: chromium:348649285
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/391240
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Kári Helgason <kthelgason@webrtc.org>
Reviewed-by: Kári Helgason <kthelgason@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44618}