[iOS] Fast-follow for Deprecated API Test Failure
In a previous CL, the RTCCameraPreviewView.m file's use of the
AVCaptureVideoOrientation API was encapsulted in a private function then
gated behind a compile-time guard. However, the invocation was not also
gated behind this same compile-time guard. This caused the WebRTC
waterfall bots to fail.

In this CL, the invocation of the private function that uses the
deprecated API is also gated behind the compile-time flag.

Change-Id: Ib94747fbe452b07b181c59387664bc1303545286
Bug: chromium:348649285
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/391560
Reviewed-by: Jeremy Leconte <jleconte@google.com>
Owners-Override: Jeremy Leconte <jleconte@google.com>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44624}