Use TransportPacketFeedback as input to TransportFeedbackDemuxer
This change refactor TransportFeedbackDemuxer to directly use the
TransportPacketFeedback from TransporFeedbackAdapter instead of
using an internal history to map a transport sequence number to
RTP packets. The purpose is to ensure TransportFeedbackDemuxer works the
same way regardless of if RFC8888 feedback or TWCC is used.

Bug: webrtc:42225697
Change-Id: I7af2dbbfab1af89e9cd25c6304ec6154430e2127
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/393341
Commit-Queue: Per Kjellander <perkj@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44762}