Add AudioMultiVector::ReadInterleavedFromIndex()
The new method differs from the existing method with the same name in
two respects:

* It reads into an InterleavedView<> instead of pointer.
* It either reads exactly the requested number of samples or fails

After examining the behavior in NetEq, where this method is called from,
it seems that what we need is a method that reads exactly the required
number of samples, not partially.

A follow-up change to NetEqImpl and SyncBuffer will change
SyncBuffer::GetNextAudioInterleaved() to read the requested buffer size
or fail, as well as update NetEqImpl::GetAudioInternal accordingly.

Bug: chromium:335805780
Change-Id: Ia78f888d52c9acb4b53b7042b31b8b0c53199e7b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/392143
Reviewed-by: Per Åhgren <peah@webrtc.org>
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44787}