Add support for setting CSRCs on audio and video senders
With this change, CSRCs can be added to video packets sent via
RTPSenderVideo::SendEncodedImage. This is implemented by keeping a list
of CSRCs in the calling class RtpVideoSender, which is included in all
calls to SendEncodedImage.

This CL is part of a chain, with the next being
https://webrtc-review.googlesource.com/c/src/+/392961. Ultimately, the
point is to support setting the CSRC list via RtpEncodingParameters.
This is done in https://webrtc-review.googlesource.com/c/src/+/392980.

Bug: b/410811496
Change-Id: I2b9c430c6b19b423f2f29cf8e81b04ad04c2b915
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/392940
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Commit-Queue: Helmer Nylén <helmern@google.com>
Cr-Commit-Position: refs/heads/main@{#44824}