Support setting a list of CSRCs in RtpEncodingParameters
CSRCs are commonly used to indicate which participants have contributed
to media packet. This change makes it possible to specify the CSRCs for
both audio and video tracks.

As demonstrated in the integration test, the CSRCs can be set either
when adding a track to a peer connection or via a SetParameters call on
an existing track.

This is the final CL in a chain that also includes
https://webrtc-review.googlesource.com/c/src/+/392940 and
https://webrtc-review.googlesource.com/c/src/+/392961.

Bug: b/410811496
Change-Id: Id0b80d22ef7e31313d43edd92b390546e376f88b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/392980
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Helmer Nylén <helmern@google.com>
Cr-Commit-Position: refs/heads/main@{#44826}