Replace use of IceCandidateInterface with IceCandidate
This is a no-code-change CL that just updates the type name in the
webrtc codebase to IceCandidate. The previous type,
IceCandidateInterface continues to exist as an alias for IceCandidate.

No-Iwyu: Lots of files touched but just updating the type name.
Bug: webrtc:406795492
Change-Id: I8dc3f532f34e5b3812ad4267336a4a60634ea01d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/395601
Auto-Submit: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44865}