Add support for setting CSRCs on audio and video senders
This is a modified version of
https://webrtc-review.googlesource.com/c/src/+/392940 to avoid breaking
downstream dependencies.

With this change, CSRCs can be added to video packets sent via
RTPSenderVideo::SendEncodedImage. This is implemented by keeping a list
of CSRCs in the calling class RtpVideoSender, which is included in all
calls to SendEncodedImage. Similarly, a list of CSRCs for audio packets
is kept on ChannelSend and added to calls to SendRtpAudio. CSRCs are
also propagated to frame transformers for audio and video.

Bug: b/410811496
Change-Id: I728934f8c190120211672e2d6dc5940bc8f83838
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/395301
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Commit-Queue: Helmer Nylén <helmern@google.com>
Cr-Commit-Position: refs/heads/main@{#44907}