Support setting a list of CSRCs in RtpEncodingParameters
This is a reupload of
https://webrtc-review.googlesource.com/c/src/+/392980.

CSRCs are commonly used to indicate which participants have contributed
to a media packet. This change makes it possible to specify the CSRCs
for both audio and video tracks.

As demonstrated in the integration test, the CSRCs can be set either
when adding a track to a peer connection or via a SetParameters call on
an existing sender.

Bug: b/410811496
Change-Id: I3f8f0b1e22680abb7d83ec27ac8025414dcacaae
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/395800
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Helmer Nylén <helmern@google.com>
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44909}