Change AddCandidate to always set the sdp_mid property
This is a follow-up to:
https://webrtc-review.googlesource.com/c/src/+/394560/13..19/api/jsep_ice_candidate.cc#b69

When JsepSessionDescription::AddCandidate is called, the sdp_mid is
sometimes provided as empty and instead an index is used.
That can cause the internal list of candidates to not have the sdp_mid()
property consistently set. This CL changes this so that the added
candidate will always have the sdp_mid() property set even if the
candidate was added by only supplying an index.

Bug: none
Change-Id: I92d1e9242a2b5eaecb0ac6a388e2a2e0d57c0cfb
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/395282
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Auto-Submit: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#44866}